likelihood update
Likelihood is four dimensional from four temperature points along the experimental peak. The posterior is exactly what to expect from a Bayesian update.  That is, the posterior is a mixture of the prior, a peak at a higher temperature, and the experimental update, a peak at a lower temperature.  The posterior model is in between the two.