Doing part of work for making ip class general
Eric, I have made a bunch of changes to run_me.py to make it easier for you to generalize the ip class. as you said, it was a bit more difficult than I expected, but I've now done the difficult parts.

Please update it the rest of the way. your "temp_points" code needs to be done on the UserInput side, since that has nothing to do with the ipclass.

After you've finished moving everything into UserInput.py that needs to go into there,

def likelihood is going to look a lot simpler.  It's going to look something like this:

def likelihood(self,discreteParameterVector):
    simulationOutput = self.UserInput.simulationFunctionWrapper(discreteParameterVector)
    simulatedResponses = self.UserInput.simulationOutputProcessingFunction(simulationOutput)
    probability_metric = multivariate_normal.pdf(x=simulatedResponses,mean=observedResponses,cov=self.UserInput.responseUncertantiesCov)
    return probability_metric, simulatedResponses

Clearly, if the code looks like above it will be a super generalized IP class! 
Don't worry about generalizing the mechanism, just worry about generalizing your class functions (which should not be too hard if you follow the FIXMEs I have left behind).

Just move things over to the UserInput slowly so that everything specific to this model is out of the class functions. So you just need to change run_me.py and UserInput.py  

Once you've done as much as you can or get stuck, let me know.