Two response example
The two response example is coded and run, and the posterior scatterplot figure added.  The example set up is there are two random variables with prior means of 1 and 5.  Each have a prior variance of 1, with no correlation between them.  Then, the first variable observes the first data point with at value of '2'.  Simultaneously, the second variable observes the second data point with a value of '3'.  The posterior mean and variance of both variables, visible in 'scatter_matrix_posterior.png', are verifiable with the analytical equations (10)-(11) of the arXiv paper (https://arxiv.org/pdf/1909.03861.pdf).  The design matrix, X, in these equations, is [1, 1], because both variables are observed.  One possible conception of this problem would be two snails, each in a separate tube.  Each of the two random variables is the location of each snail along the length of its individual tube.  Theta_1 is for the location of the first snail in tube #1, and theat_2 is the location for the second snail in tube #2.  Then, the tubes are opened and the snail's true location is observed once.  In tube #1, the location of the first snail is '2' (any length units), and in tube #2, the location of the second snail is 3.  The posterior is the updated estimate of where the snails are located after the tubes are closed again and unobservable.