Fixed multiple responses counting and plotting
It was a much bigger hassle than expected. I could not get the plotting to work for Example07f without making a few special if statements in the code. I am not very happy with the current solution, but so far it is working. It may need to be changed again once there is a 'better' multiple response example.