Attempting Unit Tests
All the unit tests are still passing except for the Example 13 unit tests.  They are **not** failing due to lack of matching output, they are failing due to an error where they are trying to take the length of an object that does not have a length, related to the response_observed_uncertainties.  Will try to debug that next.