Built wrapper for convergence diagnostic
This commit is large. The wrapper is apart of the PE_object while the conv function is not. In the conv function, some logic needed to be changed to work appropriately. The window_indices_act is now based off refined samples which should be more accurate than using the flattened postburninsamples to get sample count. All self calls are moved to the wrapper.